+++ R/utils-tidy-eval.R
@@ -42,8 +42,8 @@
 #' @name tidyeval
 #' @keywords internal
 #' @importFrom rlang expr enquo enquos sym syms .data := as_name as_label
-#' @aliases expr enquo enquos sym syms .data := as_name as_label
-#' @export expr enquo enquos sym syms .data := as_name as_label
+#' @aliases expr enquo enquos sym syms .data as_label
+#' @export expr enquo enquos sym syms .data as_label
 NULL